refactor(core): move managed entity fields to separate interface
