+++ app/scripts/modules/core/src/domain/IManagedEntity.ts
@@ -1,0 +1,34 @@
+import { IMoniker } from 'core/naming';
+
+export enum ManagedResourceStatus {
+  ACTUATING = 'ACTUATING',
+  CREATED = 'CREATED',
+  DIFF = 'DIFF',
+  ERROR = 'ERROR',
+  HAPPY = 'HAPPY',
+  PAUSED = 'PAUSED',
+  RESUMED = 'RESUMED',
+  UNHAPPY = 'UNHAPPY',
+  UNKNOWN = 'UNKNOWN',
+}
+
+export interface IManagedResourceSummary {
+  id: string;
+  kind: string;
+  status: ManagedResourceStatus;
+  moniker: IMoniker;
+  locations: {
+    account: string;
+    regions: Array<{ name: string }>;
+  };
+}
+
+export interface IManagedApplicationSummary {
+  hasManagedResources: boolean;
+  resources: IManagedResourceSummary[];
+}
+
+export interface IManagedResource {
+  managedResourceSummary?: IManagedResourceSummary;
+  isManaged?: boolean;
+}
