+++ app/scripts/modules/core/src/loadBalancer/filter/LoadBalancerFilterService.spec.ts
@@ -1,8 +1,7 @@
 import { Application } from 'core/application/application.model';
 import { ApplicationModelBuilder } from 'core/application/applicationModel.builder';
-import { ILoadBalancer, IServerGroup, ILoadBalancerGroup } from 'core/domain';
+import { ILoadBalancer, IServerGroup, ILoadBalancerGroup, IManagedResourceSummary } from 'core/domain';
 import { LoadBalancerState } from 'core/state';
-import { IManagedResourceSummary } from 'core/managed';
 
 // Most of this logic has been moved to filter.model.service.js, so these act more as integration tests
 describe('Service: loadBalancerFilterService', function() {
